+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added per-field misspellings
 - Made `exclude` option work with match all
 - Added `case_sensitive` option
+- Added `stem` option
 
 ## 3.1.0